requirements.txt
torch==2.2.1+cu121
torchvision==0.17.1+cu121
seqeval
numpy
transformers
pillow
spacy
tqdm